+++ lib/dictosaurus.dart
@@ -65,3 +65,5 @@ export 'src/dictosaurus/_index.dart'
     show DictoSaurus, DictoSaurusMixin, DictoSaurusBase;
 export 'src/_common/_index.dart' show TermProperty, PartOfSpeech;
 export 'package:gmconsult_dart_core/dart_core.dart' show Language;
+export 'package:gmconsult_dart_core/extensions.dart'
+    show LanguageStringExtensions;